Implement Evidence

+++ types/vote_set.go
@@ -162,7 +162,7 @@ func (voteSet *VoteSet) addVote(vote *Vote) (added bool, err error) {
 	}
 
 	// Add vote and get conflicting vote if any
-	added, conflicting := voteSet.addVerifiedVote(vote, blockKey, val.VotingPower)
+	added, conflicting := voteSet.addVerifiedVote(vote, blockKey, int64(val.VotingPower))
 	if conflicting != nil {
 		return added, NewConflictingVoteError(val, conflicting, vote)
 	}
